Rename files and make all functions public.
